+++ src/dbtarzan/config/password/CommonJsonProtocols.scala
@@ -1,9 +1,9 @@
-package dbtarzan.config
+package dbtarzan.config.password
 
 import spray.json._
 
 object PasswordJsonProtocol extends DefaultJsonProtocol {
-  import dbtarzan.config.Password
+  import dbtarzan.config.password.Password
   implicit object PasswordFormat extends JsonFormat[Password] {
     def write(password: Password) = JsString(password.key)
     def read(json: JsValue): Password = json match {
