+++ src/dbtarzan/config/password/PasswordEncryption.scala
@@ -1,4 +1,4 @@
-package dbtarzan.config
+package dbtarzan.config.password
 
 import dbtarzan.config.util.{ Encryption, ByteArrayHex }
