+++ src/dbtarzan/db/actor/DatabaseWorker.scala
@@ -8,7 +8,7 @@ import scala.collection.mutable.HashMap
 
 import dbtarzan.db.util.ExceptionToText
 import dbtarzan.config.connections.ConnectionData
-import dbtarzan.config.EncryptionKey
+import dbtarzan.config.password.EncryptionKey
 import dbtarzan.db._
 import dbtarzan.messages._
 import dbtarzan.localization.Localization
